Refactor action registry helpers into dedicated modules
